services/auth-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3
motor==3.3.2
redis==5.0.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
httpx==0.26.0
aiokafka==0.10.0
numpy==1.26.3
